Restructure the deeplink artifact
1. Simplified the base package from "deeplink.parseIntent.singleModule" to "deeplink.basic".
2. Added `ui` and `deeplinkutil` packages to separate sample ui code from parsing/matching helpers
3. Separate classes within DeepLinkUtil into their own separate files
4. Separate DeepLinkRequest into two separate classes
     - DeepLinkRequest: parse uri and store parse result
     - DeepLinkMatcher: takes a request + pattern and matches the two
5. Change type <T> upper bound to NavKey to make the helpers more general

Other misc non-structural changes:
1. Added kdocs in CreateDeepLinkActivity to explain how this recipe works
2. inlined String parsers into #getTypeParser
